moved cart functions from App.js file to Cart.js file and updated the Shop and Product files. add to cart buttons are not working yet
