+++ src/Components/pages/cart/Cart.js
@@ -1,11 +1,34 @@
 import React from 'react';
+import { useState } from 'react';
 
-const Cart = (props) => {
-  const { cartItems, onAdd, onRemove } = props;
-  const itemPrice = cartItems.reduce((a, c) => a + c.price * c.qty, 0);
-  const taxPrice = itemPrice * 0.056;
-  const shippingPrice = itemPrice > 2000 ? 0 : 50;
-  const orderTotal = itemPrice + taxPrice + shippingPrice;
+const Cart = () => {
+  const [cartItems, setCartItems] = useState([]);
+
+  const onAdd = (product) => {
+    const exist = cartItems.find((x) => x.id === product.id);
+    if (exist) {
+      setCartItems(
+        cartItems.map((x) =>
+          x.id === product.id ? { ...exist, qty: exist.qty + 1 } : x
+        )
+      );
+    } else {
+      setCartItems([...cartItems, { ...product, qty: 1 }]);
+    }
+  };
+  const onRemove = (product) => {
+    const exist = cartItems.find((x) => x.id === product.id);
+    if (exist.qty === 1) {
+      setCartItems(cartItems.filter((x) => x.id !== product.id));
+    } else {
+      setCartItems(
+        cartItems.map((x) =>
+          x.id === product.id ? { ...exist, qty: exist.qty - 1 } : x
+        )
+      );
+    }
+  };
+  const orderTotal = cartItems.reduce((a, c) => a + c.price * c.qty, 0);
     
     return (
         <div>
@@ -31,18 +54,6 @@ const Cart = (props) => {
                     <div>
                         <hr></hr>
                         <div>
-                            <div>Item Price</div>
-                            <div>${itemPrice.toFixed(2)}</div>
-                        </div>
-                        <div>
-                            <div>Tax Price</div>
-                            <div>${taxPrice.toFixed(2)}</div>
-                        </div>
-                        <div>
-                            <div>Shipping</div>
-                            <div>${shippingPrice.toFixed(2)}</div>
-                        </div>
-                        <div>
                             <div>
                                 <strong>Order Total</strong>
                             </div>
@@ -64,3 +75,15 @@ const Cart = (props) => {
 
 export default Cart;
 
+// <div>
+// <div>Item Price</div>
+// <div>${itemPrice.toFixed(2)}</div>
+// </div>
+// <div>
+// <div>Tax Price</div>
+// <div>${taxPrice.toFixed(2)}</div>
+// </div>
+// <div>
+// <div>Shipping</div>
+// <div>${shippingPrice.toFixed(2)}</div>
+// </div>
