+++ src/Components/pages/store/Product.js
@@ -1,14 +1,29 @@
 import React from 'react';
+import { useState } from 'react';
 
 const Product = (props) => {
-    const { product, onAdd } = props;
+    const { product } = props;
+    const [cartItems, setCartItems] = useState([]);
+
+    const onAdd = (product) => {
+        const exist = cartItems.find((x) => x.id === product.id);
+        if (exist) {
+        setCartItems(
+            cartItems.map((x) =>
+            x.id === product.id ? { ...exist, qty: exist.qty + 1 } : x
+            )
+        );
+        } else {
+        setCartItems([...cartItems, { ...product, qty: 1 }]);
+        }
+    };
     return (
         <div>
             <img className="small" src={product.image} alt={product.name}></img>
             <h3>{product.name}</h3>
             <div>${product.price}</div>
             <div>
-            <button onClick={() => onAdd(product)}>Add to Cart</button>
+            <button onClick={onAdd}>Add to Cart</button>
             </div>
         </div>
     );
