+++ src/Components/pages/store/Shop.js
@@ -1,8 +1,10 @@
 import React from "react";
+import data from "../../../data";
 import Product from "./Product";
 
 const Shop = (props) => {
-    const { products, onAdd } = props;
+    const {products} = data;
+    const { onAdd } = props;
     return(
         <div className="shop">
             <h2>Products</h2>
